Set up Sass and organize.
