+++ v2/package.json
@@ -6,15 +6,23 @@
     "@types/ace": "^0.0.39",
     "ace-builds": "^1.3.3",
     "asciidoctor.js": "^1.5.7-beta.1",
+    "bootstrap-sass": "^3.3.7",
+    "jquery-ui": "^1.12.1",
     "mobx": "^5.0.2",
     "mobx-react": "^5.2.3",
+    "node-sass-chokidar": "^1.3.3",
     "react": "^16.4.1",
+    "react-bootstrap": "^0.32.2",
     "react-dom": "^16.4.1",
+    "react-fontawesome": "^1.6.1",
     "react-scripts-ts": "2.16.0"
   },
   "scripts": {
-    "start": "react-scripts-ts start",
-    "build": "react-scripts-ts build",
+    "build-css": "node-sass-chokidar --include-path ./node_modules --include-path ./src/ src/ -o src/",
+    "watch-css": "npm run build-css && node-sass-chokidar --include-path ./node_modules --include-path ./src src/ -o src/ --watch --recursive",
+    "start-js": "react-scripts-ts start",
+    "start": "npm-run-all -p watch-css start-js",
+    "build": "npm run build-css && react-scripts-ts build",
     "test": "react-scripts-ts test --env=jsdom",
     "eject": "react-scripts-ts eject"
   },
@@ -23,6 +31,7 @@
     "@types/node": "^10.3.3",
     "@types/react": "^16.3.17",
     "@types/react-dom": "^16.0.6",
+    "npm-run-all": "^4.1.3",
     "typescript": "^2.9.2"
   }
 }
