+++ v2/public/fonts/bootstrap
@@ -1,0 +1,1 @@
+../../node_modules/bootstrap-sass/assets/fonts/bootstrap
