+++ v2/src/app/_base.scss
@@ -1,0 +1,93 @@
+html,
+body {
+  height: 100%;
+}
+
+.navbar {
+  border: none;
+}
+
+$navbar-brand-img-height: 14px;
+.navbar-brand > img {
+  height: $navbar-brand-img-height;
+  margin-top: (
+    ($navbar-height - $navbar-brand-img-height) / 2 - $navbar-padding-vertical
+  );
+}
+
+.navbar-toggle {
+  color: #fff;
+}
+
+.navbar #header-menu {
+  .save-icon {
+    width: 1em;
+    height: auto;
+    vertical-align: -35%;
+
+    @media (min-width: $grid-float-breakpoint) {
+      margin-left: 0.2em;
+    }
+    @media (max-width: $grid-float-breakpoint) {
+      margin-right: 0.3em;
+    }
+  }
+
+  li.dropdown > a {
+    text-transform: uppercase;
+    letter-spacing: 1px;
+
+    @media (min-width: $grid-float-breakpoint) {
+      span.icon-label {
+        display: none;
+      }
+    }
+  }
+  li.dropdown-header {
+    text-transform: uppercase;
+    letter-spacing: 1.25px;
+
+    @media (max-width: $grid-float-breakpoint) {
+      color: #fff;
+      opacity: 0.85;
+    }
+  }
+
+  ul.dropdown-menu .fa {
+    @media (min-width: $grid-float-breakpoint) {
+      color: $gray;
+    }
+
+    &.fa-toggle-on {
+      color: $brand-success;
+    }
+    &.fa-toggle-off {
+      @media (min-width: $grid-float-breakpoint) {
+        color: $gray;
+      }
+    }
+  }
+
+  ul.dropdown-menu .open-local {
+    position: relative;
+    padding: 0;
+
+    & > label {
+      font-weight: normal;
+      margin: 0;
+      padding: 3px 20px; // Same as regular li > a padding
+      width: 100%;
+      cursor: pointer;
+    }
+    & > input[type='file'] {
+      position: absolute;
+      opacity: 0;
+      z-index: -1;
+    }
+  }
+}
+
+#loading-banner {
+  padding-top: ($navbar-height * 3);
+  text-align: center;
+}
