+++ v2/src/app/_bootstrap-vars.scss
@@ -1,0 +1,13 @@
+$navbar-border-radius: 0;
+$navbar-inverse-bg: #444;
+$navbar-inverse-toggle-border-color: transparent;
+$navbar-inverse-toggle-hover-bg: transparent;
+$navbar-inverse-link-disabled-color: #fff;
+$navbar-inverse-border: transparent;
+$navbar-inverse-color: #fdfdfd;
+$navbar-inverse-link-color: #fdfdfd;
+
+$navbar-height-collapsed: 54px;
+
+// TODO
+$icon-font-path: '/';
