+++ v2/src/app/app.scss
@@ -1,0 +1,4 @@
+@import 'app/bootstrap-vars';
+@import 'bootstrap-sass/assets/stylesheets/bootstrap';
+
+@import 'app/base';
