+++ v2/src/app/app.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import App from './App';
+import App from './app';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
