+++ v2/src/app/app.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import './app.css';
+
+class App extends React.Component {
+  public render() {
+    // TODO
+    return <div />;
+  }
+}
+
+export default App;
